requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pymongo>=4.6.0
pytesseract>=0.3.10
httpx>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0
Pillow>=10.2.0
requests>=2.31.0
